etcd: implement Create and Update functions (#24)
This commit introduces the function `put`, which is a common function
that all *Create* and `Update` functions - i.e. `CreateNs` or
`UpdateEndp` will use, as the operation of inserting and updating
objects are both performed with a *put* operation.
Finally, `Create` and `Update` functions are now implemented for all
objects.

Signed-off-by: Elis Lulja <elulja@cisco.com>