Include etcd settings in the operator settings (#24)
The operator yaml settings now have a `serviceRegistry` field that
contains settings about the service registry to use.
`etcd` is a new field in the operator's YAML settings that includes a
list of endpoints of the etcd cluster to connect to.
Additionally, this commit moves service directory settings from gcloud
to serviceRegistry to be more consistent and deprecates the old way.

Signed-off-by: Elis Lulja <elulja@cisco.com>