Better way to clean up when exiting (#24)
This commit takes care of cleaning everything up.
The previous method, `os.Exit` exited the main goroutine without
honoring `defer` calls - this is due to kube-builder original code:
this means that the context would not be canceled and, with etcd, the
client would not be closed, potentially
exposing leaky goroutines.
While this is not very critical when run in a pod, it's still a better
choice to clean up our own stuff.

Signed-off-by: Elis Lulja <elulja@cisco.com>